Make sure credentials sign in works
