+++ pages/api/auth/[...nextauth].ts
@@ -17,11 +17,22 @@ let url = new URL('https://www.dropbox.com/oauth2/authorize');
 url.searchParams.set('token_access_type', 'offline');
 url.searchParams.set('scope', ['account_info.read', 'files.content.read'].join(' '));
 
+function profile(account: users.FullAccount) {
+  return {
+    id: account.account_id,
+    email: account.email,
+    name: account.name.display_name,
+    image: account.profile_photo_url,
+    pathRoot: account.root_info.root_namespace_id,
+  };
+}
+
 const dropbox = DropboxProvider({
   id: 'dropbox',
   clientId: CLIENT_ID,
   clientSecret: CLIENT_SECRET,
   authorization: url.toString(),
+  profile,
   userinfo: {
     async request(context) {
       let headers = new Headers();
@@ -31,15 +42,6 @@ const dropbox = DropboxProvider({
       return res.json();
     },
   },
-  profile(account: users.FullAccount) {
-    return {
-      id: account.account_id,
-      email: account.email,
-      name: account.name.display_name,
-      image: account.profile_photo_url,
-      pathRoot: account.root_info.root_namespace_id,
-    };
-  },
 });
 
 const credentials = CredentialsProvider({
@@ -48,8 +50,22 @@ const credentials = CredentialsProvider({
   credentials: {
     refresh_token: { label: 'Token', type: 'password' },
   },
-  async authorize(credentials, req) {
-    throw new Error('');
+  async authorize(credentials) {
+    try {
+      if (credentials?.refresh_token == null) return null;
+
+      let { accessToken, expiresAt } = await fetchFreshAccessToken(credentials.refresh_token);
+      let account = await fetchCurrentAccount(accessToken);
+
+      return {
+        ...profile(account),
+        access_token: accessToken,
+        expires_at: expiresAt,
+        refresh_token: credentials.refresh_token,
+      };
+    } catch (error) {
+      throw new Error('CredentialsSignin');
+    }
   },
 });
 
@@ -57,19 +73,24 @@ const options: AuthOptions = {
   providers: [dropbox, credentials],
   pages: {
     signIn: config['route.signin'],
+    error: config['route.signin'],
   },
   callbacks: {
     async jwt({ token, account, user }) {
       if (account != null && user != null) {
-        token.accessToken = ensure(account.access_token, 'No access token received from auth endpoint');
-        token.refreshToken = ensure(account.refresh_token, 'No refresh token received from auth endpoint');
-        token.pathRoot = ensure(user.pathRoot, 'No path root received from user info');
-        /**
-         * For some reason, can't really find out how, next-auth is giving back an expiry date that is multiplied by
-         * 1000 too many times. Usually it should come back as seconds from the provider, so multiplying once should be
-         * okey. But somehow by here the value has been multiplied again.
-         */
-        token.expiresAt = ensure(account.expires_at, 'No expiry received from auth endpoint') / 1000 + Date.now();
+        try {
+          token.accessToken = ensure(account.access_token, 'No access token received from auth endpoint');
+          token.refreshToken = ensure(account.refresh_token, 'No refresh token received from auth endpoint');
+          token.pathRoot = ensure(user.pathRoot, 'No path root received from user info');
+          /**
+           * For some reason, can't really find out how, next-auth is giving back an expiry date that is multiplied by
+           * 1000 too many times. Usually it should come back as seconds from the provider, so multiplying once should be
+           * okey. But somehow by here the value has been multiplied again.
+           */
+          token.expiresAt = ensure(account.expires_at, 'No expiry received from auth endpoint') / 1000 + Date.now();
+        } catch (error) {
+          throw error;
+        }
       }
 
       if (Date.now() < token.expiresAt) return token;
@@ -80,7 +101,6 @@ const options: AuthOptions = {
       session.refreshToken = token.refreshToken;
       session.expiresAt = token.expiresAt;
       session.pathRoot = token.pathRoot;
-      session.error = token.error;
 
       return session;
     },
@@ -105,31 +125,37 @@ const RefreshToken = z.object({
 
 async function refreshAccessToken(token: JWT): Promise<JWT> {
   try {
-    let url = new URL('https://api.dropbox.com/oauth2/token');
-    url.searchParams.set('grant_type', 'refresh_token');
-    url.searchParams.set('refresh_token', token.refreshToken);
-    url.searchParams.set('client_id', CLIENT_ID);
-
-    const response = await fetch(url, {
-      headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-      method: 'POST',
-    });
-    if (!response.ok) throw response;
-
-    const refreshedTokens = RefreshToken.parse(await response.json());
-
-    let expiresAt = Date.now() + refreshedTokens.expires_in * 1000;
-    return {
-      ...token,
-      accessToken: refreshedTokens.access_token,
-      expiresAt,
-    };
+    let fresh = await fetchFreshAccessToken(token.refreshToken);
+    return { ...token, ...fresh };
   } catch (error) {
-    return {
-      ...token,
-      error: 'RefreshAccessTokenError',
-    };
+    throw new Error('RefreshAccessTokenError');
   }
+}
+
+async function fetchFreshAccessToken(refreshToken: string) {
+  let url = new URL('https://api.dropbox.com/oauth2/token');
+  url.searchParams.set('grant_type', 'refresh_token');
+  url.searchParams.set('refresh_token', refreshToken);
+  url.searchParams.set('client_id', CLIENT_ID);
+
+  const response = await fetch(url, {
+    headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+    method: 'POST',
+  });
+  if (!response.ok) throw response;
+
+  const refreshedTokens = RefreshToken.parse(await response.json());
+
+  let expiresAt = Date.now() + refreshedTokens.expires_in * 1000;
+  return { accessToken: refreshedTokens.access_token, expiresAt };
+}
+
+async function fetchCurrentAccount(accessToken: string) {
+  let headers = new Headers();
+  headers.set('Authorization', `Bearer ${accessToken}`);
+
+  let res = await fetch('https://api.dropboxapi.com/2/users/get_current_account', { method: 'POST', headers });
+  return res.json() as Promise<users.FullAccount>;
 }
 
 declare module 'next-auth' {
@@ -148,7 +174,6 @@ declare module 'next-auth' {
     refreshToken: string;
     expiresAt: number;
     pathRoot: string;
-    error?: 'RefreshAccessTokenError';
     user: User;
   }
 }
@@ -159,6 +184,5 @@ declare module 'next-auth/jwt' {
     refreshToken: string;
     expiresAt: number;
     pathRoot: string;
-    error?: 'RefreshAccessTokenError';
   }
 }
